Refactor Template Binary Sensor
