+++ src/panels/config/entities/editor-tabs/settings/entity-settings-helper-tab.ts
@@ -43,10 +43,10 @@ import {
   updateInputText,
 } from "../../../../../data/input_text";
 import {
-  deleteBinarySensor,
-  fetchBinarySensor,
-  updateBinarySensor,
-} from "../../../../../data/binary_sensor";
+  deleteTemplateBinarySensor,
+  fetchTemplateBinarySensor,
+  updateTemplateBinarySensor,
+} from "../../../../../data/template.binary_sensor";
 import { showConfirmationDialog } from "../../../../../dialogs/generic/show-dialog-box";
 import type { HomeAssistant } from "../../../../../types";
 import type { Helper } from "../../../helpers/const";
@@ -85,9 +85,9 @@ const HELPERS = {
     delete: deleteInputSelect,
   },
   binary_sensor: {
-    fetch: fetchBinarySensor,
-    update: updateBinarySensor,
-    delete: deleteBinarySensor,
+    fetch: fetchTemplateBinarySensor,
+    update: updateTemplateBinarySensor,
+    delete: deleteTemplateBinarySensor,
   },
 };
